Add web worker compatibility
