+++ .gitignore
@@ -14,6 +14,4 @@ artifacts
 tmp
 package-lock.json
 /*.html
-/js/src/resolc.*
-/js/dist/
 /build
