+++ js/examples/node/resolc.js
@@ -1,0 +1,1 @@
+../../../target/wasm32-unknown-emscripten/release/resolc.js
