+++ js/examples/web/resolc.wasm
@@ -1,0 +1,1 @@
+../../../target/wasm32-unknown-emscripten/release/resolc.wasm
